interface: input and output arrays to methods via file descriptors
Add a new feature for passing array arguments to and from methods
using a unix pipe. The pipe file descriptors are transferred via dbus
along with the normal parameters. The feature is selected by using
'A' instead of 'a' for arrays in the xml interface description. This
transfer mode only supports simple types and cannot be nested.