interface: make async_call_ready variable local to Proxy function instead of a private member
This variable beeing a private member might cause dead locks if the proxy function is called asyncronously from multiple threads. Making it local to the function should prevent that, because each thread would have its own version of that variable on the thread-local stack.
Also, an error message is created if a function that expects file descriptors does not get any file descriptors.