saftbus: remove Socket Class to simplify the code
the Socket class was not doing so much anymore. it was only encapsulating a file descriptor.
So I removed it. Unfortunately saftd segfaulted as some point which means I probably introduced
a bug. I will fix bugs after doing all the simplifications because I think it will be easier to
fix bugs in simpler code.