saftd: in multithreaded applications with high signal rates it may happen that the signal pipes get full
Previously, this caused everything to crash. The obvious solution to just make the pipe blocking doesnt work because it can lead to deadlocks in multithreaded applications: saftd tries to send a signal into a blocked pipe to a proxy, and the same proxy is blocked because saftd doesnt reply to the socket. This commit fixes this by buffering the signals on the saftd side and schedule a resend callback if the signal pipe is blocked. This implementation is naive and slow. The faster implemetation doesnt work yet.