burst-ctl: write the user commands to a dedicated buffer in the shared memory
The common-libs library defines its buffer structure in the shared memory.
Some of these buffers are used to communicate with the host system.
The library does not use the mailbox slot.
Therefore, the user commands must be written to a dedicated command buffer
in the shared memory. On user command completion, LM32 responds through
a subscribed mailbox slot.