fixed a bug in the removal of saftbus::Sources from saftbus::Loop
Problem was that Source* were used as handles for later removal. But that does not
work because Sources can remove themselves from the Loop by returning false in the
dispatch method. When later a Source* handle is used to explicitely remove a
Source, there might alredy be a new Soure at this memory location. In such a case
a completely unrelated Source is removed from the Loop causing all kind of strange
behavior.
To fix this, a proper source handle is introduced that saves unique id of Source
and Loop instance that it refers to.