saft-io-ctl works now
signal fds are now only closed on the server side when the client quits.
They are not closed if their use count goes to 0, because later the client
might want to reuse this connection.

An alternative to this would be that the Proxy constructor checks if the
signal connection is still acitve (check for POLLHUP) and re establish it
(make a new socketpair and send one end to server) if it is not active anymore