new idea for dealing with the modular firmware problem
The idea of adding firmware drivers to a TimingReceiver is much too narrow and limits possibilities of developers.
Instead it should be possible to extend the functionality of a TimingReceiver - while using its ressources - in a
more general way. It should be possible to add so called TimingReceiverAddons to the TimingReceiver object.
These may be firmware drivers, interacting with one ore many LM32 cores, but can also do something completetly
differen. For example take control over several IOs and bit-bang a serial protocol. The addon mechanism has no limits.
It only provides a pointer to a TimingReceiver object to the TimingReceiverAddon constructor. The only thing any Addon
needs to do is provide a list of Interfaces and the ObjectPaths such that a remote user can instanciate the services
over saftubs proxies.