saft-*: Handle empty device list for `-f`flag
If the saftd service does not have any devices attached,
`Saftd_Proxy::create()->getDevices()` will return an empty map.
This change checks for this edge case before trying to create a
`TimingReceiverProxy` which will fail with `std::bad_alloc` otherwise.
With this change an error message is printed and an error code of -1 is
returned, which matches the behaviour of calling `saft-* <nonexistent
device> ...`.