fix saftbus::write_all
the problem is that write_all is blocking. Under some conditions this can cause
a deadlock with a client process. One can not simply make write_all non-blocking
by aborting an incomplete data transfer because the client has no means to check
data integrety and relies on completed transfer. The implemented solution buffers
the remaining data in case the socket is not ready to receive more. If the function
is called again on the same file descriptor, the buffered data (if any is present)
is sent first. In this way, clients will always receive complete buffers without
the danger of causing a deadlock.